test: adapt unit tests after merge
