+++ frontend/test/unit/stores/accounts.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi, type Mock } from 'vitest';
 import { setActivePinia, createPinia } from 'pinia';
 import { useAccountsStore } from '@/stores';
-import { useWallet, useShortAddress } from '@/hooks';
+import { useWallet } from '@/hooks';
 
 const testKey1 =
   '0x4104e41989791c1e1f3a889710d8c5509b98ee62e5fabced8a3f538f5afd392b'; // ! NEVER USE THIS PRIVATE KEY
@@ -16,7 +16,7 @@ vi.mock('@/hooks', () => ({
 
 describe('useAccountsStore', () => {
   let accountsStore: ReturnType<typeof useAccountsStore>;
-  const mockWallet = {
+  const mockWeb3 = {
     privateKeyToAccount: vi.fn(),
     generatePrivateKey: vi.fn(),
   };
@@ -26,8 +26,10 @@ describe('useAccountsStore', () => {
 
   beforeEach(() => {
     setActivePinia(createPinia());
-    (useWallet as Mock).mockReturnValue(mockWallet);
-    (useShortAddress as Mock).mockReturnValue(mockShortAddress);
+    (useWallet as Mock).mockReturnValue({
+      web3: mockWeb3,
+      shortenAddress: mockShortAddress.shortenAddress,
+    });
 
     // Mock localStorage
     vi.stubGlobal('localStorage', {
@@ -38,8 +40,8 @@ describe('useAccountsStore', () => {
 
     accountsStore = useAccountsStore();
 
-    mockWallet.privateKeyToAccount.mockClear();
-    mockWallet.generatePrivateKey.mockClear();
+    mockWeb3.privateKeyToAccount.mockClear();
+    mockWeb3.generatePrivateKey.mockClear();
     mockShortAddress.shortenAddress.mockClear();
     (localStorage.getItem as Mock).mockClear();
     (localStorage.getItem as Mock).mockClear();
@@ -48,7 +50,7 @@ describe('useAccountsStore', () => {
 
   it('should generate a new account', () => {
     const newPrivateKey = '0xnewkey';
-    mockWallet.generatePrivateKey.mockReturnValue(newPrivateKey);
+    mockWeb3.generatePrivateKey.mockReturnValue(newPrivateKey);
     const generatedKey = accountsStore.generateNewAccount();
 
     expect(generatedKey).toBe(newPrivateKey);
@@ -77,7 +79,7 @@ describe('useAccountsStore', () => {
   it('should handle errors in displayAddress computation', () => {
     accountsStore.currentPrivateKey = '0xinvalidkey';
 
-    mockWallet.privateKeyToAccount.mockImplementation(() => {
+    mockWeb3.privateKeyToAccount.mockImplementation(() => {
       throw new Error('Invalid private key');
     });
 
@@ -100,10 +102,10 @@ describe('useAccountsStore', () => {
     const privateKey = testKey1;
     const account = { address: testAddress1 };
     accountsStore.currentPrivateKey = privateKey;
-    mockWallet.privateKeyToAccount.mockReturnValue(account);
+    mockWeb3.privateKeyToAccount.mockReturnValue(account);
 
     expect(accountsStore.currentUserAddress).toBe(testAddress1);
-    expect(mockWallet.privateKeyToAccount).toHaveBeenCalledWith(privateKey);
+    expect(mockWeb3.privateKeyToAccount).toHaveBeenCalledWith(privateKey);
   });
 
   it('should return an empty string for currentUserAddress when no private key is set', () => {
